+++ packages/bodiless-components/__tests__/withFilterByTags.test.tsx
@@ -35,6 +35,7 @@ const getMockNode = () => {
   const getters = {
     getNode: jest.fn((path: string[]) => testTags[path[path.length - 1]]),
     getKeys: jest.fn(() => Object.keys(testTags)),
+    hasError: jest.fn(),
   };
   const actions = {
     setNode: jest.fn(),