+++ packages/bodiless-core-ui/src/GlobalContextMenu.tsx
@@ -27,12 +27,13 @@ import {
   ComponentFormError, ComponentFormSubmitButton, ComponentFormList, ComponentFormListItem,
   ComponentFormDescription, ComponentFormWarning, ComponentFormFieldWrapper,
   ComponentFormFieldTitle, ComponentFormCheckBox, ComponentFormRadio, ComponentFormRadioGroup,
-  ComponentFormSelect, ComponentFormOption, ComponentFormTextArea,
+  ComponentFormSelect, ComponentFormOption, ComponentFormTextArea, ContextSubMenu,
+  ToolbarButtonLabel,
 } from '@bodiless/ui';
 import ReactTagsField from './ReactTags';
 
 const Toolbar = flow(
-  addClasses('bl-flex bl-flex-col bl-bg-black bl-rounded bl-z-50 bl-p-grid-2 bl-fixed bl-top-grid-0 bl-left-grid-0 bl-text-white'),
+  addClasses('bl-flex bl-flex-col bl-bg-black bl-rounded bl-z-50 bl-p-grid-1 bl-fixed bl-top-grid-0 bl-left-grid-0 bl-text-white'),
   addProps({ role: 'toolbar', 'aria-label': 'Global Context Menu Left', id: 'global-context-menu' }),
 )(Div);
 
@@ -94,9 +95,11 @@ const ui: ContextMenuUI = {
   ComponentFormWarning,
   ComponentFormList,
   ComponentFormListItem,
+  ContextSubMenu,
   Icon,
   Toolbar,
   ToolbarButton,
+  ToolbarButtonLabel,
   ToolbarDivider,
   FormWrapper,
   Tooltip: GlobalTooltip,