+++ packages/bodiless-core/src/ActivateContext.tsx
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-import * as React from 'react';
+import React from 'react';
 import { useContextActivator } from './hooks';
 
 type ActivateOnEffectState = {