+++ packages/bodiless-core/src/ContextMenu/StructuredChildren.tsx
@@ -1,0 +1,73 @@
+import React, {
+  FC, ReactElement, ReactNode, ComponentType,
+} from 'react';
+import { flow } from 'lodash';
+import { ContextMenuGroupProps } from '../Types/ContextMenuTypes';
+
+type GroupTree = {
+  [name: string]: {
+    element: ReactElement,
+    members: GroupTree,
+  }
+};
+
+export const asElementArray = (children: ReactNode): ReactElement[] => React.Children
+  .toArray(children)
+  .filter(React.isValidElement);
+
+export const addMissingGroups = (GroupComponent: ComponentType<ContextMenuGroupProps>) => (
+  (elements: ReactElement[]): ReactElement[] => elements.reduce(
+    (acc: ReactElement[], el: ReactElement) => {
+      if (el.props.group && !acc.find(el$ => el$.props.name === el.props.group)) {
+        return [...acc, <GroupComponent name={el.props.group} />];
+      }
+      return acc;
+    },
+    elements,
+  )
+);
+
+export const buildGroupTree = (elements: ReactElement[], groupName: string = '_default'): GroupTree => elements
+  .filter(el => (el.props.group || '_default') === groupName)
+  .reduce((acc, child) => ({
+    ...acc,
+    [child.props.name]: {
+      element: child,
+      members: buildGroupTree(elements, child.props.name),
+    },
+  }), {});
+
+export const cloneChildren = (props: any = {}) => (tree: GroupTree): ReactElement[] => Object
+  .getOwnPropertyNames(tree)
+  .reduce((acc: ReactElement[], name: string) => {
+    const entry = tree[name];
+    const newElement = React.cloneElement(entry.element, {
+      ...props,
+      key: entry.element.props.name,
+      children: cloneChildren(props)(entry.members),
+    });
+    return [...acc, newElement];
+  }, []);
+
+const buildChildren = (
+  DefaultGroupComponent: ComponentType<ContextMenuGroupProps> = React.Fragment,
+  props: any = {},
+) => flow(
+  asElementArray,
+  addMissingGroups(DefaultGroupComponent),
+  buildGroupTree,
+  cloneChildren(props),
+);
+
+type GroupedChildrenProps = {
+  components: {
+    Group: ComponentType<ContextMenuGroupProps>,
+  },
+  [prop: string]: any,
+};
+
+const StructuredChildren: FC<GroupedChildrenProps> = ({ components, children, ...rest }) => (
+  <>{buildChildren(components.Group, rest)(children)}</>
+);
+
+export default StructuredChildren;